feat: accomodate BTC-timestamping (#40)
This PR closes #4.

The [change](https://github.com/babylonlabs-io/babylon/pull/23) in
Babylon removed voting power from a finality provider if it has no
timestamped pub rand for a certain height. Therefore, from the finality
provider side, as long as it has voting power, the existence of a
timestamped pub rand is ensured. No need to do an extra query to check
it. This PR accommodates this change