fix: fpd panics on jailed finality provider (#78)
In the previous implementation, once a finality provider is jailed, when
the fpd is restarted, it will panic due to error from Babylon when fast
sync. This PR fixed this issue with the following changes:
1. We implemented `unjail` finality provider CLI other than using the
one from Babylon. This implementation will set the fp instance status to
`inactive` after `unjail` tx is successfully sent (will later updated to
active if the fp has voting power).
2. Once `jailed` error is detected while starting a fp instance, it will
fail but the fpd will not panic, meaning that the loop for updating
stored fp status continues running.

Now the flow of jailing/unjailing becomes the follows:
1. fpd detects jailing via err when sending a fp sig or a loop for
checking fp status
2. once `jailed` detected, fpd terminates the fp instance without
terminating the program
3. the operator checks fp signing info to get the `jail_until` via
`babylond q finality signing-info [fp-pk-hex]`
4. after the `jail_until` is passed, the operator can unjail the fp by
executing `fpd unjail-finality-provider [fp-pk-hex]`
5. if everything goes well, the fp will continue sending finality votes
if it has voting power after a period of waiting for state transition