chore: Remove sync fp status loop (#208)
Closes #191. In particular, this PR:
- removed the loop for syncing fp status but syncing in the main thread
during fp app initiation.
- the loop is no longer needed as we have decided no fp will be
registered via chain upgrade
- we still need syncing status during initiation in case the fp is
slashed or jailed
- in `monitorStatusUpdate`, the check for jailing and slashing is
removed. This is because the removal of voting power due to slashing or
jailing happens in the next block. Countering jailing and slashing in
the loop will cause the fp to not vote when it still have voting power
for the current block