chore: Refactor `CommitPubRand` (#233)
Closes #225. The major goal of this pr is to refactor `CommitPubRand`
but turned out included many small refactorings along the way (sorry for
the mixup). Notable changes are:
- Split `retryCommitPubRandUntilBlockFinalized` into
`ShouldCommitRandomness` and `CommitPubRand`. The former checks whether
a new commit should be made based on the last committed height and the
tip height. If `true`, it also returns the start height of the commit.
The latter makes a new commit based on the start height returned by the
former.
- Rename `config.MinRandHeightGap` to `config.TimestampingDelayBlocks`
which is an estimated value of the latency between a commit is submitted
and it is timestamped by the btc-timestamping protocol
- Remove status update loop. Instead, the status update is done when
querying voting power during vote submission in `ProcessBlocksToVote`

Reviewers, please take a careful look at `fp_instance.go`, especially
how `ShouldCommitRandomness` determines the start height with
consideration of btc-timestamping delay