chore: Refactor to start from the last finalized height (#257)
Closes https://github.com/babylonlabs-io/finality-provider/issues/253.

The poller should start from the `lastFinalizedHeight` other than
`lastVotedHeight`. This is to deal with cases where the gap between the
two heights is large so that the delay of catching up could cause
jailing.

This change will not give up many rewards because rewards are not
assigned to late votes due to
[ADR](https://github.com/babylonlabs-io/pm/blob/main/adr/adr-039-finality-voting-rewarding.md).