fix: no add failed cycles count when chain poller no found new blocks (#327)
When chain poller try to got new blocks, it will count a `failedCycles`,
when it > `maxFailedCycles`, the finality-provider will exit by:

```go
		if failedCycles > maxFailedCycles {
			cp.logger.Fatal("the poller has reached the max failed cycles, exiting")
		}
```

BUT in chain poller, when the poller no got new blocks during
`PollInterval`, it will also add `failedCycles`.

So when the chain no produce blocks in `PollInterval` *
`maxFailedCycles`, the fp will exit.

For some chain, it may no produce blocks when no new tx, and in some
times, the chain will pause by chain upgrade, so it will make the fp
exit by the error.

So in this PR, we make `failedCycles ` no update when no new blocks.