chore: Modify default config and add validation (#418)
The point of this PR is to ensure that users cannot insert config values
that will cause the fp to crash or jailed. Notable changes are:

- Remove config items NumPubRandMax, TimestampingDelayBlocks,
RandomnessCommitInterval but hardcode them as they are critical ones
that should not be changeable. TimestampingDelayBlocks is calculated
based on w=300 and babylon block production time of 10s
- Add validations to config items. Note that MinPubRand to ensure that a
commit can pass verification on Babylon side.

---------

Co-authored-by: samricotta <37125168+samricotta@users.noreply.github.com>
Co-authored-by: Vitalis Salis <vitsalis@babylonlabs.io>