feat(rollup): rollup BSN FP setup (#518)
First step of
https://github.com/babylonlabs-io/finality-provider/issues/507

This PR creates the vanilla setup for the rollup BSN finality provider.
This is a major refactoring that includes:

- [x] creating a `bsn/rollup-finality-provider` folder fole rollup BSN
FP
- [x] migrating rollup client controller to the folder
- [x] creating dedicated config for rollup BSN FP
- [x] migrating e2e tests for rollup BSN FP 
- [x] fixing Makefile

This is only the first step and `rollup-fpd` binary is not functional
yet. The necessary next step is to fix all the command line arguments
(https://github.com/babylonlabs-io/finality-provider/pull/517)