feat: batch eots sign (#715)
Instead of requesting eots to sign height by height, we do it in
batches.

Local modified (not committed) cosmos (as we have 1s block time) e2e
catch-up test shows that for a batch of 60, execution time for old
implementations is `1.149s` and for the new implementation `95.10ms`.
These times should be even more exaggerated as the local e2e test has
minimal network latency.